requirements.txt
fastmcp
openai
aiofiles
python-dotenv
